Remove WhiteNoise
We’re now serving assets from S3 when running on PaaS. So we’ll set
the appropriate caching headers there or in Cloudfront. This means that
the app no longer needs to serve cache headers, which is what we were
using WhiteNoise for.

This commits removes WhiteNoise in favour of letting Flask handle the
serving of static assets.